Done atm and pos extraction. IB data missiong from switch report

+++ helper.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 class helper:
     def tagFinder(self, dlo_ite, token):
         for each in dlo_ite:
@@ -22,7 +21,7 @@ class helper:
             elif each.tag == 'Currency':
                 dic['Currency'] = each.text
             elif each.tag == 'Amount':
-                dic['Amount'] = Decimal(each.text)
+                dic['Amount'] = float(each.text)
             else:
                 continue
         return dic
